# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пользователь успешно зарегистрирован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Пользователь с таким email уже существует"
          }
        },
        "summary": "Регистрация нового пользователя",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный вход",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Неверные учетные данные"
          }
        },
        "summary": "Вход в систему",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/forgot": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Если пользователь существует, отправлено письмо для сброса пароля",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordResponseDto"
                }
              }
            }
          }
        },
        "summary": "Запрос на сброс пароля",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Токены успешно обновлены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Неверный или истекший refresh token"
          }
        },
        "summary": "Обновление токенов",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/ai/improve-text": {
      "post": {
        "operationId": "AiController_improveText",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImproveTextDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Текст улучшен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImproveTextResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Улучшить текст с помощью AI",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/cover-letter": {
      "post": {
        "operationId": "AiController_generateCoverLetter",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoverLetterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Сопроводительное письмо сгенерировано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverLetterResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Сгенерировать сопроводительное письмо с помощью AI",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/analyze-vs-job": {
      "post": {
        "operationId": "AiController_analyzeVsJob",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeVsJobDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Анализ выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeVsJobResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Проанализировать резюме относительно описания вакансии",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/achievements": {
      "get": {
        "operationId": "AchievementsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список достижений",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AchievementResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить все достижения текущего пользователя",
        "tags": [
          "achievements"
        ]
      },
      "post": {
        "operationId": "AchievementsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAchievementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Достижение успешно создано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AchievementResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Создать новое достижение",
        "tags": [
          "achievements"
        ]
      }
    },
    "/api/achievements/{id}": {
      "get": {
        "operationId": "AchievementsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID достижения",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о достижении",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AchievementResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Достижение не найдено"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить достижение по ID",
        "tags": [
          "achievements"
        ]
      },
      "put": {
        "operationId": "AchievementsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID достижения",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAchievementDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Достижение обновлено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AchievementResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Достижение не найдено"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обновить достижение",
        "tags": [
          "achievements"
        ]
      },
      "delete": {
        "operationId": "AchievementsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID достижения",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Достижение удалено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AchievementResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен"
          },
          "404": {
            "description": "Достижение не найдено"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить достижение",
        "tags": [
          "achievements"
        ]
      }
    },
...
  },
  "info": {
    "title": "PortfolioX API",
    "description": "Backend API для платформы PortfolioX - совместное резюме + портфолио",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email пользователя",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя (минимум 8 символов)",
            "example": "password123",
            "minLength": 8
          },
          "name": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "Иван Иванов"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID пользователя",
            "example": "clx1234567890"
          },
          "email": {
            "type": "string",
            "description": "Email пользователя",
            "example": "user@example.com"
          },
          "emailVerified": {
            "type": "boolean",
            "description": "Подтвержден ли email",
            "example": false
          },
          "name": {
            "type": "object",
            "description": "Имя пользователя",
            "nullable": true
          },
          "headline": {
            "type": "object",
            "description": "Заголовок профиля",
            "nullable": true
          },
          "avatarUrl": {
            "type": "object",
            "description": "URL аватара",
            "nullable": true
          },
          "publicHandle": {
            "type": "object",
            "description": "Публичный handle",
            "nullable": true
          },
          "locale": {
            "type": "object",
            "description": "Локаль",
            "example": "en",
            "nullable": true
          },
          "profiles": {
            "description": "Профили пользователя",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileResponseDto"
            }
          },
          "_count": {
            "description": "Количество элементов",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserCountsDto"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата обновления",
            "example": "2025-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "email",
          "emailVerified",
          "createdAt",
          "updatedAt"
        ]
      },
      "SignupResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "description": "Данные пользователя",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserResponseDto"
              }
            ]
          }
        },
        "required": [
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email пользователя",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "description": "Данные пользователя",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserResponseDto"
              }
            ]
          }
        },
        "required": [
          "user"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email пользователя для сброса пароля",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ForgotPasswordResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Сообщение о статусе запроса",
            "example": "If user exists, password reset email sent"
          }
        },
        "required": [
          "message"
        ]
      },
      "RefreshResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "description": "Данные пользователя",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserResponseDto"
              }
            ]
          }
        },
        "required": [
          "user"
        ]
      },
...
      "LinkResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID ссылки",
            "example": "clx1234567890"
          },
          "url": {
            "type": "string",
            "description": "URL ссылки",
            "example": "https://github.com/user"
          },
          "title": {
            "type": "object",
            "description": "Название ссылки",
            "example": "GitHub",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2025-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "url",
          "createdAt"
        ]
      },
      "ProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID профиля",
            "example": "clx1234567890"
          },
          "userId": {
            "type": "string",
            "description": "ID пользователя",
            "example": "clx1234567890"
          },
          "bio": {
            "type": "object",
            "description": "Биография",
            "nullable": true
          },
          "visible": {
            "type": "boolean",
            "description": "Видимость профиля",
            "example": true
          },
          "links": {
            "description": "Ссылки профиля",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LinkResponseDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата обновления",
            "example": "2025-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "userId",
          "visible",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserCountsDto": {
        "type": "object",
        "properties": {
          "resumes": {
            "type": "number",
            "description": "Количество резюме",
            "example": 3
          },
          "projects": {
            "type": "number",
            "description": "Количество проектов",
            "example": 5
          },
          "achievements": {
            "type": "number",
            "description": "Количество достижений",
            "example": 2
          }
        },
        "required": [
          "resumes",
          "projects",
          "achievements"
        ]
      },
...
      "ImproveTextDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Текст для улучшения",
            "example": "Я разработал приложение"
          },
          "goal": {
            "type": "string",
            "description": "Цель улучшения текста",
            "example": "Сделать текст более профессиональным"
          },
          "language": {
            "type": "string",
            "description": "Язык текста",
            "example": "ru"
          }
        },
        "required": [
          "text"
        ]
      },
      "ImproveTextResponseDto": {
        "type": "object",
        "properties": {
          "improvedText": {
            "type": "string",
            "description": "Улучшенный текст",
            "example": "Разработал и внедрил систему, которая увеличила производительность на 40%"
          }
        },
        "required": [
          "improvedText"
        ]
      },
      "CoverLetterDto": {
        "type": "object",
        "properties": {
          "resumeId": {
            "type": "string",
            "description": "ID резюме",
            "example": "clx1234567890"
          },
          "jobDescription": {
            "type": "string",
            "description": "Описание вакансии",
            "example": "Ищем опытного разработчика с опытом работы..."
          }
        },
        "required": [
          "resumeId",
          "jobDescription"
        ]
      },
      "CoverLetterResponseDto": {
        "type": "object",
        "properties": {
          "coverLetter": {
            "type": "string",
            "description": "Сгенерированное сопроводительное письмо",
            "example": "Уважаемый [Имя],\n\nЯ пишу, чтобы выразить заинтересованность..."
          }
        },
        "required": [
          "coverLetter"
        ]
      },
      "AnalyzeVsJobDto": {
        "type": "object",
        "properties": {
          "resumeText": {
            "type": "string",
            "description": "Текст резюме",
            "example": "Опытный разработчик с 5+ годами опыта..."
          },
          "jobDescription": {
            "type": "string",
            "description": "Описание вакансии",
            "example": "Ищем опытного разработчика с опытом работы..."
          }
        },
        "required": [
          "resumeText",
          "jobDescription"
        ]
      },
      "AnalyzeVsJobResponseDto": {
        "type": "object",
        "properties": {
          "score": {
            "type": "number",
            "description": "Оценка совместимости резюме с вакансией (от 0 до 1)",
            "example": 0.75
          },
          "suggestions": {
            "description": "Предложения по улучшению резюме",
            "example": [
              "Добавьте больше информации о опыте работы с React",
              "Укажите конкретные метрики достижений"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "strengths": {
            "description": "Сильные стороны резюме",
            "example": [
              "Отличный опыт работы с Node.js",
              "Сильные навыки работы в команде"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "missingKeywords": {
            "description": "Отсутствующие ключевые слова",
            "example": [
              "TypeScript",
              "Docker",
              "Kubernetes"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "score",
          "suggestions",
          "strengths",
          "missingKeywords"
        ]
      },
      "AchievementResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID достижения",
            "example": "clx1234567890"
          },
          "userId": {
            "type": "string",
            "description": "ID пользователя",
            "example": "clx1234567890"
          },
          "title": {
            "type": "string",
            "description": "Название достижения",
            "example": "Победитель хакатона"
          },
          "body": {
            "type": "object",
            "description": "Описание достижения",
            "nullable": true
          },
          "date": {
            "type": "object",
            "description": "Дата достижения",
            "example": "2025-01-01T00:00:00.000Z",
            "nullable": true
          },
          "link": {
            "type": "object",
            "description": "Ссылка на достижение",
            "example": "https://example.com/achievement",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата создания",
            "example": "2025-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Дата обновления",
            "example": "2025-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "userId",
          "title",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateAchievementDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Название достижения",
            "example": "Победитель хакатона 2024"
          },
          "body": {
            "type": "string",
            "description": "Описание достижения",
            "example": "Занял первое место на хакатоне по разработке мобильных приложений"
          },
          "date": {
            "type": "string",
            "description": "Дата достижения",
            "example": "2025-01-15"
          },
          "link": {
            "type": "string",
            "description": "Ссылка на достижение",
            "example": "https://example.com/achievement"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateAchievementDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Название достижения",
            "example": "Победитель хакатона 2024"
          },
          "body": {
            "type": "string",
            "description": "Описание достижения",
            "example": "Занял первое место на хакатоне по разработке мобильных приложений"
          },
          "date": {
            "type": "string",
            "description": "Дата достижения",
            "example": "2025-01-15"
          },
          "link": {
            "type": "string",
            "description": "Ссылка на достижение",
            "example": "https://example.com/achievement"
          }
        }
      },
...
    }
  }
}